Rename database table for discarded recommendations (CONDEC_DISC_REC)

+++ src/main/java/de/uhd/ifi/se/decision/management/jira/persistence/tables/DecisionGroupInDatabase.java
@@ -2,6 +2,7 @@ package de.uhd.ifi.se.decision.management.jira.persistence.tables;
 
 import java.sql.SQLException;
 
+import de.uhd.ifi.se.decision.management.jira.model.DocumentationLocation;
 import de.uhd.ifi.se.decision.management.jira.model.KnowledgeElement;
 import net.java.ao.RawEntity;
 import net.java.ao.schema.AutoIncrement;
@@ -16,7 +17,8 @@ import net.java.ao.schema.Table;
  * @see KnowledgeElement
  */
 @Table("CondecDecGroup")
-public interface DecisionGroupInDatabase extends RawEntity<Integer> {
+public interface DecisionGroupInDatabase extends RawEntity<Long> {
+
 	@AutoIncrement
 	@PrimaryKey("ID")
 	long getId();
@@ -27,12 +29,28 @@ public interface DecisionGroupInDatabase extends RawEntity<Integer> {
 
 	void setGroup(String group);
 
+	/**
+	 * @return id of the {@link KnowledgeElement} that this group is assigned to.
+	 */
 	long getSourceId();
 
+	/**
+	 * @param id
+	 *            of the {@link KnowledgeElement} that this group is assigned to.
+	 */
 	void setSourceId(long id);
 
+	/**
+	 * @return {@link DocumentationLocation#getIdentifier()} of the
+	 *         {@link KnowledgeElement} that this group is assigned to.
+	 */
 	String getSourceDocumentationLocation();
 
+	/**
+	 * @param documentationLocation
+	 *            {@link DocumentationLocation#getIdentifier()} of the
+	 *            {@link KnowledgeElement} that this group is assigned to.
+	 */
 	void setSourceDocumentationLocation(String documentationLocation);
 
 	String getProjectKey();
@@ -40,11 +58,11 @@ public interface DecisionGroupInDatabase extends RawEntity<Integer> {
 	void setProjectKey(String projectKey);
 
 	/**
-	 * Deletes the {@group DecisionGroupInDatabase } object, i.e., removes it from
+	 * Deletes the {@link DecisionGroupInDatabase} object, i.e., removes it from
 	 * database.
-	 * 
-	 * @param groupToDelete
-	 *            {@group DecisionGroupInDatabase} object.
+	 *
+	 * @param groupAssignmentToDelete
+	 *            {@link DecisionGroupInDatabase} object.
 	 * @return true if deletion was successful, false otherwise.
 	 */
 	static boolean deleteGroup(DecisionGroupInDatabase groupAssignmentToDelete) {
